+++ src/pages/cms/[id]/[cid]/[mid]/index.tsx
@@ -2,7 +2,7 @@ import BottomTab from '@components/bottom-tab';
 import { Wrapper } from '@styles/styles';
 import { useRouter } from 'next/router';
 import { Card, Text, Spacer, Input, Button, Checkbox } from '@geist-ui/react';
-import { createOption, getOption } from '@/apis/store/store';
+import { createOption, getOption } from '@/apis/cms-store/store';
 import { useState } from 'react';
 import { StoreContainer, StyledText } from '@/styles/cmsStyles';
 import { Option } from '@/types/type';
